Fix dependency on zlib on windows, and calls to ceil()
git-svn-id: https://svn.osgeo.org/gdal/trunk@15660 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/georaster/georaster_priv.h
@@ -45,7 +45,7 @@ class GeoRasterWrapper;
 //  ---------------------------------------------------------------------------
 
 #define CALCULATEBLOCK( bn, xo, yo, bb, tc, tr ) \
-    ( ( (int) ceil( ( bn - 1 ) / bb ) * tc * tr ) + ( yo * tc ) + xo )
+    ( ( (int) ceil( (double) ( ( bn - 1 ) / bb ) ) * tc * tr ) + ( yo * tc ) + xo )
 
 //  ---------------------------------------------------------------------------
 //  Link between bands+pyramid levels and modified <shrinked> georaster object
